+++ src/collective/frontpage/profiles/default/types/Section.xml
@@ -94,7 +94,7 @@
   <action
       action_id="edit"
       category="object"
-      condition_expr="not:object/@@plone_lock_info/is_locked_for_current_user|python:True"
+      condition_expr=""
       i18n:attributes="title"
       title="Edit"
       url_expr="string:${object_url}/edit"
